style: move useRRWebEventStream to clickhouse file
